metodos put y delete

+++ src/main/java/com/ciclo3/reto/retogrupog24/entidad/Category.java
@@ -1,6 +1,7 @@
 package com.ciclo3.reto.retogrupog24.entidad;
 
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -16,7 +17,7 @@ import java.util.List;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String name;
     private String description;
     @OneToMany(cascade = {CascadeType.PERSIST},mappedBy = "category")
